+++ GeneticSolution/GeneticSolution/Container.cs
@@ -1,7 +1,6 @@
 namespace GeneticSolution;
 
-public record Container(int width, int height, int depth, Tuple<int, int, int> coordinates,
-    Tuple<int, int, int>? potentialCoordinates = null)
+public record Container(int width, int height, int depth)
 {
     public int Volume => width * height * depth;
 }
