+++ GeneticSolution/GeneticSolution/DBLF.cs
@@ -1,0 +1,13 @@
+namespace GeneticSolution;
+
+public class DBLF
+{
+    public DBLF(Point currentPosition, Point possiblePosition)
+    {
+        CurrentPosition = currentPosition;
+        PossiblePosition = possiblePosition;
+    }
+
+    public Point CurrentPosition { get; }
+    public Point PossiblePosition { get; }
+}
