+++ GeneticSolution/GeneticSolution/Point.cs
@@ -1,0 +1,17 @@
+namespace GeneticSolution;
+
+public class Point
+{
+    public Point(int x, int y, int z)
+    {
+        X = x;
+        Y = y;
+        Z = z;
+    }
+
+    public int X { get; }
+    public int Y { get; }
+    public int Z { get; }
+
+    public int Volume => X * Y * Z;
+}
