+++ GeneticSolution/GeneticSolution/Rotations.cs
@@ -1,0 +1,11 @@
+namespace GeneticSolution;
+
+public enum Rotations
+{
+    LWH = 0,
+    WLH = 1,
+    LHW = 2,
+    HLW = 3,
+    HWL = 4,
+    WHL = 5
+}
